code: Created Card component
