+++ src/components/SnippetComponents/Snippets/DropDowns.js
@@ -1,7 +1,7 @@
 export const dropDownSnippets = [
   {
     title: "Simple Dropdown Snippet",
-    jsxCode: `(props) => (
+    jsxCode: `
         <div className="flex items-center space-x-4">
           <select
             className="border border-gray-300 rounded-md px-3 py-2 focus:outline-none"
@@ -13,7 +13,7 @@ export const dropDownSnippets = [
             <option value="option3">Option 3</option>
           </select>
         </div>
-      )`,
+      `,
     cssCode: `<div style="border: 1px solid #ccc; padding: 10px; border-radius: 5px;">
         <select style="border: 1px solid #ccc; padding: 8px; border-radius: 4px;">
           <option value="">Select an option...</option>
@@ -25,7 +25,7 @@ export const dropDownSnippets = [
   },
   {
     title: "Grouped Dropdown Snippet",
-    jsxCode: `(props) => (
+    jsxCode: `
         <div className="flex items-center space-x-4">
           <select
             className="border border-gray-300 rounded-md px-3 py-2 focus:outline-none"
@@ -41,7 +41,7 @@ export const dropDownSnippets = [
             </optgroup>
           </select>
         </div>
-      )`,
+      `,
     cssCode: `<div style="border: 1px solid #ccc; padding: 10px; border-radius: 5px;">
         <select style="border: 1px solid #ccc; padding: 8px; border-radius: 4px;">
           <optgroup label="Group 1">
@@ -57,7 +57,7 @@ export const dropDownSnippets = [
   },
   {
     title: "Custom Styled Dropdown Snippet",
-    jsxCode: `(props) => (
+    jsxCode: `
         <div className="flex items-center space-x-4">
           <select
             className="border border-purple-500 rounded-lg px-4 py-2 bg-purple-50 text-purple-900 shadow-md focus:outline-none"
@@ -69,7 +69,7 @@ export const dropDownSnippets = [
             <option value="option3">Option 3</option>
           </select>
         </div>
-      )`,
+      `,
     cssCode: `<div style="border: 1px solid #ccc; padding: 10px; border-radius: 5px;">
         <select style="border: 1px solid #ccc; padding: 8px; border-radius: 4px; background-color: #F3E5F5; color: #6A1B9A;">
           <option value="">Select an option...</option>
